+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/EffectedVideoChatActivity.kt
@@ -93,6 +93,13 @@ class EffectedVideoChatActivity : AppCompatActivity() {
                 "EMBOSS" -> {
                     addGPUImageFilter(GPUImageEmbossFilter())
                 }
+                "SEPIA" -> {
+                    addGPUImageFilter(GPUImageSepiaFilter())
+                }
+                "NONE" -> {
+                    // For Debug
+                    addGPUImageFilter(GPUImageFilter())
+                }
                 else -> {}
             }